chore: extract properties to a separate file
